Registration system shifted to spring.

+++ src/main/java/com/owo/OwoDokan/service/admin_related/order/Shop_keeper_order_service.java
@@ -12,6 +12,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.util.Optional;
+
 @Service
 public class Shop_keeper_order_service {
     private final Order_repo order_repo;
@@ -26,55 +28,55 @@ public class Shop_keeper_order_service {
 
     public ResponseEntity addOrder(Shop_keeper_orders shop_keeper_order_param, String mobile_number) {
 
-        Shops shops;
+        Optional<Shops> shopsOptional = shopRepository.getByPhone(mobile_number);
 
-        try
+        if(shopsOptional.isPresent())
         {
-            shops = shopRepository.getByPhone(mobile_number);
-        }catch (Exception e)
-        {
-            return new ResponseEntity(HttpStatus.FAILED_DEPENDENCY);
+            Shops shops = shopsOptional.get();
+
+            Shop_keeper_orders shopKeeperOrders = new Shop_keeper_orders();
+
+            shopKeeperOrders.setAdditional_comments(shop_keeper_order_param.getAdditional_comments());
+            shopKeeperOrders.setCoupon_discount(shop_keeper_order_param.getCoupon_discount());
+            shopKeeperOrders.setDate(shop_keeper_order_param.getDate());
+            shopKeeperOrders.setDelivery_address(shop_keeper_order_param.getDelivery_address());
+            shopKeeperOrders.setMethod(shop_keeper_order_param.getMethod());
+            shopKeeperOrders.setReceiver_phone(shop_keeper_order_param.getReceiver_phone());
+            shopKeeperOrders.setShop_phone(shop_keeper_order_param.getShop_phone());
+            shopKeeperOrders.setShipping_state(shop_keeper_order_param.getShipping_state());
+            shopKeeperOrders.setTime_slot(shop_keeper_order_param.getTime_slot());
+            shopKeeperOrders.setOrder_time(shop_keeper_order_param.getOrder_time());
+            shopKeeperOrders.setTotal_amount(shop_keeper_order_param.getTotal_amount());
+
+            shopKeeperOrders.setShops(shops);
+            shops.getShopKeeperOrders().add(shopKeeperOrders);
+
+            for(Shop_keeper_ordered_products shop_keeper_ordered_products : shop_keeper_order_param.getShop_keeper_ordered_products())
+            {
+                shop_keeper_ordered_products.setShop_keeper_orders(shopKeeperOrders);
+                shopKeeperOrders.getShop_keeper_ordered_products().add(shop_keeper_ordered_products);
+            }
+
+            try
+            {
+                shopRepository.save(shops);
+            }catch(Exception e)
+            {
+                return new ResponseEntity(HttpStatus.FAILED_DEPENDENCY);
+            }
+
+            try
+            {
+                cartRepo.deleteById(shopKeeperOrders.getShop_phone());
+                return new ResponseEntity(HttpStatus.CREATED);
+            }catch (Exception e)
+            {
+                return new ResponseEntity(HttpStatus.FAILED_DEPENDENCY);
+            }
         }
-
-
-        Shop_keeper_orders shopKeeperOrders = new Shop_keeper_orders();
-
-        shopKeeperOrders.setAdditional_comments(shop_keeper_order_param.getAdditional_comments());
-        shopKeeperOrders.setCoupon_discount(shop_keeper_order_param.getCoupon_discount());
-        shopKeeperOrders.setDate(shop_keeper_order_param.getDate());
-        shopKeeperOrders.setDelivery_address(shop_keeper_order_param.getDelivery_address());
-        shopKeeperOrders.setMethod(shop_keeper_order_param.getMethod());
-        shopKeeperOrders.setReceiver_phone(shop_keeper_order_param.getReceiver_phone());
-        shopKeeperOrders.setShop_phone(shop_keeper_order_param.getShop_phone());
-        shopKeeperOrders.setShipping_state(shop_keeper_order_param.getShipping_state());
-        shopKeeperOrders.setTime_slot(shop_keeper_order_param.getTime_slot());
-        shopKeeperOrders.setOrder_time(shop_keeper_order_param.getOrder_time());
-        shopKeeperOrders.setTotal_amount(shop_keeper_order_param.getTotal_amount());
-
-        shopKeeperOrders.setShops(shops);
-        shops.getShopKeeperOrders().add(shopKeeperOrders);
-
-        for(Shop_keeper_ordered_products shop_keeper_ordered_products : shop_keeper_order_param.getShop_keeper_ordered_products())
+        else
         {
-            shop_keeper_ordered_products.setShop_keeper_orders(shopKeeperOrders);
-            shopKeeperOrders.getShop_keeper_ordered_products().add(shop_keeper_ordered_products);
-        }
-
-        try
-        {
-            shopRepository.save(shops);
-        }catch(Exception e)
-        {
-            return new ResponseEntity(HttpStatus.FAILED_DEPENDENCY);
-        }
-
-        try
-        {
-            cartRepo.deleteById(shopKeeperOrders.getShop_phone());
-            return new ResponseEntity(HttpStatus.CREATED);
-        }catch (Exception e)
-        {
-            return new ResponseEntity(HttpStatus.FAILED_DEPENDENCY);
+            return ResponseEntity.status(HttpStatus.FAILED_DEPENDENCY).body("Can not find shop");
         }
     }
